src/requirements.txt
celery[redis]
django
django-admin-interface
django-celery-beat
django-debug-toolbar
django-money[exchange]
django-structlog[celery]
psycopg2-binary
python-dotenv
requests
tailslide
